\name{==.SBML}
\alias{==.SBML}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Check the equality of the species and reactions of two SBML models}
\description{
  This function tests the equivalence of two models with respect to the species and reaction data frames generated by summary.  
}
\usage{
"==.SBML"(model1, model2, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model1}{The first of the two model objects of class SBML which are to be compared.}
  \item{model2}{The second model object.}
  \item{...}{ For compatibility with \code{==} of the base package.}
 }
% \details{no details}

\value{ A list containing the following two boolean dataframes 
 \item{species}{ The equality of species information tabularized  as a data frame.}
  \item{reactions}{ The equality of reaction information tabularized as a dataframe.}
}

\author{ Tomas Radivoyevitch  (radivot@hal.cwru.edu)}

\seealso{ \code{\link{summary.SBML}}  }
\examples{
library(SBMLR)  
curto1=readSBMLR(file.path(system.file(package="SBMLR"), "models/curto.r"))  
curto2=readSBML(file.path(system.file(package="SBMLR"), "models/curto.xml"))  
curto1==curto2

}
\keyword{ math }% at least one, from doc/KEYWORDS
